Avoid a warning on Windows.
SVN-Revision: 3948

+++ libarchive/archive_read_support_format_7zip.c
@@ -2489,7 +2489,7 @@ read_Header(struct _7zip *zip, struct _7z_header_info *h,
 		/* The high 16 bits of attributes is a posix file mode. */
 		entries[i].mode = entries[i].attr >> 16;
 		if (entries[i].flg & HAS_STREAM) {
-			if (sindex >= si->ss.unpack_streams)
+			if ((size_t)sindex >= si->ss.unpack_streams)
 				return (-1);
 			if (entries[i].mode == 0)
 				entries[i].mode = AE_IFREG | 0777;
